# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Bookworm API
  version: 0.1.0
  description: Mandatory paths/schemas per MASTER-BUILD-SPEC section 9. All endpoints versioned under /v1.
servers:
  - url: http://localhost:3001
...
paths:
  /v1/admin/access:
    get:
      operationId: adminAccess
      summary: Verify platform administrator access
      description: Requires a verified user listed in ADMIN_USER_IDS; workspace ownership alone is insufficient.
      responses:
        "200":
          description: Platform administrator
          content:
            application/json:
              schema: { type: object, required: [admin], properties: { admin: { type: boolean, const: true } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/users:
    get:
      operationId: adminListUsers
      summary: List user profiles and organization memberships (platform admin only)
      parameters:
        - { $ref: "#/components/parameters/AdminLimit" }
        - { $ref: "#/components/parameters/AdminOffset" }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
          description: Literal display-name substring or exact UUID.
      responses:
        "200": { description: "Object containing users, memberships, limit and offset." }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/jobs:
    get:
      operationId: adminListJobs
      summary: Inspect durable AI, publishing or manuscript-import jobs (platform admin only)
      parameters:
        - { $ref: "#/components/parameters/AdminLimit" }
        - { $ref: "#/components/parameters/AdminOffset" }
        - name: type
          in: query
          schema: { type: string, enum: [ai, publishing, document], default: ai }
        - name: status
          in: query
          schema: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
      responses:
        "200": { description: "Object containing jobs, type, limit and offset." }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/flags:
    get:
      operationId: adminListFlags
      summary: List feature flags across scopes (platform admin only)
      responses:
        "200": { description: Object containing flags ordered by key. }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/flags/{key}:
    put:
      operationId: adminSetFlag
      summary: Set a scoped feature flag and atomically record the audit event
      description: Platform administrators only. Omitted config preserves existing configuration. Null scope IDs share one unique global-scope record.
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled]
              properties:
                enabled: { type: boolean }
                scopeType: { type: string, minLength: 1, maxLength: 40, default: global }
                scopeId: { type: [string, "null"], maxLength: 120, default: null }
                config: { type: object, additionalProperties: true }
      responses:
        "200": { description: Object containing the persisted flag. }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/admin/support:
    get:
      operationId: adminListSupport
      summary: List support tickets (platform admin only)
      parameters:
        - { $ref: "#/components/parameters/AdminLimit" }
        - { $ref: "#/components/parameters/AdminOffset" }
        - name: status
          in: query
          schema: { type: string, maxLength: 40 }
      responses:
        "200": { description: "Object containing tickets, limit and offset." }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/support/{id}:
    post:
      operationId: adminUpdateTicket
      summary: Change support-ticket status or priority (platform admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, pending, resolved, closed] }
                priority: { type: string, enum: [low, normal, high, urgent] }
      responses:
        "200": { description: Object containing the updated ticket. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/audit:
    get:
      operationId: adminListAudit
      summary: Inspect audit events (platform admin only)
      parameters:
        - { $ref: "#/components/parameters/AdminLimit" }
        - { $ref: "#/components/parameters/AdminOffset" }
        - { name: entity, in: query, schema: { type: string, maxLength: 80 } }
        - { name: action, in: query, schema: { type: string, maxLength: 80 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        "200": { description: "Object containing entries, limit and offset." }
        "403": { $ref: "#/components/responses/Error" }
  /v1/admin/usage/summary:
    get:
      operationId: adminUsageSummary
      summary: Aggregate a bounded usage sample by organization and meter
      description: Platform administrators only. Samples up to 10000 events and returns up to 50 organizations. Meter quantities have different units and should be displayed separately.
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 90, default: 30 } }
      responses:
        "200": { description: "Object containing days and orgs; each org includes organizationId, total and byMeter." }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    AdminLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    AdminOffset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
...
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
security:
  - bearerAuth: []